[GN-107] rename: move providers to layouts folder
